Converter: adding annotations to track converted image
Add option `harbor_annotation: true` to append harbor specified annotations
to the manifest of converted image:

```
{
  ...
  annotation：{
    "io.goharbor.artifact.v1alpha1.acceleration.driver.name": "estargz"
    "io.goharbor.artifact.v1alpha1.acceleration.source.digest": "sha256:dbad66bcfe29ef383157a3e122acbd08cd2ebd40f5658afa2ae62c52ffe26e9f"
  }
}
```

The first annotation is used to identify different accelerated image
formats in harbor.

The second annotation is used to reference the source (original) image,
which can be used to avoid duplicate conversion or tracking the relationship
between source image and converted image in harbor portal.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>