nydus: refactor driver using converter
Use converter package from nydus-snapshotter to refactor layer
conversion workflow.

The package converts OCI tar layer to nydus blob layer, and then
merges all blob layers into final bootstrap layer.

The benefit of this change is that there is no need for containerd
to unpack the OCI layer and mount it as a snapshot, which improves
the speed and stability of the image layer conversion.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>