converter: support skipping tls verification
An error is thrown when converting image to HTTPS registry with
self-signed TLS cert:

failed to do request: x509: certificate signed by unknown authority

We currently do not provide any option for the HTTPS registry to skip
insecure self-signed certificate verification, so this commit changes
the semantics of the previous `insecure: true` option to support it.
if users specify this option, then skips verifying HTTPS certs.

And this commit allows acceld automatically fallback to plain HTTP
if the first HTTPS request fails for an HTTP registry when whether
or not the `insecure: true` option is enabled.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>