[nydus] Remove old config's reference in config descriptor's annotations when adding empty layer
Trying to push a manifest with an added empty layer to ECR currently fails because the manifest's config descriptor has an annotation field with a referrence to the old config digest. The failure is ` Invalid parameter at 'ImageManifest' failed to satisfy constraint: 'Invalid JSON syntax'`.
This field should be accepted by ECR because it's present in the spec but since there is no hard need to keep it as an annotation on the config descriptor, this commit remove the annotation entirely since the original config can still be discovered by finding the original OCI manifest using the digest stored in the new manifest annotations.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>