[adapter] Use context without cancel for GC
The GC triggered in a goroutine after the Convert operation is done was called with the same context passed down. However, when the convert action is called synchronously (accelctl --sync), the associated context is canceled when the sync operation is completed. Since the GC was run after, it was failing with a `context canceled` error.
To solve this, we now remove the `cancel` from the context passed down to the GC.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>