[nydus] Remove timestamp in empty layer history
While having a timestamp gives an interesting information on when the nydus image was built, the consequence is that every reconversion of the same image will yield a different digest because the timestamp changed which is a nightmare from a caching perspective. To avoid this, the timestamp is removed from the empty layer to ensure a consistent digest.

Signed-off-by: Baptiste Girard-Carrabin <baptiste.girardcarrabin@datadoghq.com>